%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{addColumn.BaseFileSection}
\alias{addColumn.BaseFileSection}
\alias{BaseFileSection.addColumn}
\alias{addColumn.BaseFileSection}
\alias{addColumn,BaseFileSection-method}


 \title{Adds a column to the other column names}

 \description{
   Adds a column to the other column names.
 }

 \usage{\method{addColumn}{BaseFileSection}(this, name, ...)}

 \arguments{
  \item{name}{A column name.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
