%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileParameters.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{attachParameters.BaseFileParameters}
\alias{attachParameters.BaseFileParameters}
\alias{BaseFileParameters.attachParameters}
\alias{attachParameters.BaseFileParameters}
\alias{attachParameters,BaseFileParameters-method}


 \title{Assigns the parameters of a BASE parameters section locally}

 \description{
   Assigns the parameters of a BASE parameters section locally.
 }

 \usage{\method{attachParameters}{BaseFileParameters}(this, ..., envir=parent.frame())}

 \arguments{
  \item{...}{Arguments passed to \code{\link[aroma.Base:getParameters.BaseFileParameters]{*getParameters}()}.}
  \item{envir}{The \code{\link[base]{environment}} where to assign the parameters.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileParameters}}.
 }
\keyword{internal}
\keyword{methods}
