%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileBioassaysetinfo.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{BaseFileBioassaysetinfo}
\docType{class}
\alias{BaseFileBioassaysetinfo}


 \title{The BaseFileBioassaysetinfo class extending BaseFileSection}

 \description{
  Package:  aroma.Base \cr
\bold{Class BaseFileBioassaysetinfo}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[aroma.Base]{BaseFileSection}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{BaseFileBioassaysetinfo}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{BaseFileBioassaysetinfo}\cr
extends \link[aroma.Base]{BaseFileSection}\cr


 }

 \usage{BaseFileBioassaysetinfo(...)}

 \arguments{
   \item{...}{Arguments passed to constructor of superclass.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[aroma.Base:as.character.BaseFileBioassaysetinfo]{as.character}} \tab Gets a string description of object.\cr
 \tab \code{\link[aroma.Base:getName.BaseFileBioassaysetinfo]{getName}} \tab Gets the name of the bioassay set.\cr
 \tab \code{\link[aroma.Base:validate.BaseFileBioassaysetinfo]{validate}} \tab Validates bioassay-set information section.\cr
}


 \bold{Methods inherited from BaseFileSection}:\cr
addColumn, as.character, as.list, equals, getColumns, getData, getDataByKeys, getHeader, getHeaders, getType, hasColumns, hasData, hasHeaders, nbrOfHeaders, print, setColumns, setData, setDataFieldByKeys, setHeader, setType, sizeOfData, validate

 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save

 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
