%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{equals.BaseFileSection}
\alias{equals.BaseFileSection}
\alias{BaseFileSection.equals}
\alias{equals.BaseFileSection}
\alias{equals,BaseFileSection-method}


 \title{Checks if a BASE file section equals another}

 \description{
   Checks if a BASE file section equals another.
 }

 \usage{\method{equals}{BaseFileSection}(this, other, ...)}

 \arguments{
  \item{other}{Other object.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base:logical]{TRUE}}, if they are equal, otherwise \code{\link[base:logical]{FALSE}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
