%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getAllDataFiles.BaseFile}
\alias{getAllDataFiles.BaseFile}
\alias{BaseFile.getAllDataFiles}
\alias{getAllDataFiles.BaseFile}
\alias{getAllDataFiles,BaseFile-method}


 \title{Gets the names of a external data files used by a BASE file}

 \description{
   Gets the names of a external data files used by a BASE file.
 }

 \usage{\method{getAllDataFiles}{BaseFile}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFile}}.
 }
\keyword{internal}
\keyword{methods}
