%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getAnnotationColumns.BaseFileAssays}
\alias{getAnnotationColumns.BaseFileAssays}
\alias{BaseFileAssays.getAnnotationColumns}
\alias{getAnnotationColumns.BaseFileAssays}
\alias{getAnnotationColumns,BaseFileAssays-method}


 \title{Gets the names of the annotation columns}

 \description{
   Gets the names of the annotation columns.
 }

 \usage{\method{getAnnotationColumns}{BaseFileAssays}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:getColumns.BaseFileAssays]{*getColumns}()}
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
