%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSpots.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getAssayFields.BaseFileSpots}
\alias{getAssayFields.BaseFileSpots}
\alias{BaseFileSpots.getAssayFields}
\alias{getAssayFields.BaseFileSpots}
\alias{getAssayFields,BaseFileSpots-method}


 \title{Gets the names of the assay-field data columns}

 \description{
   Gets the names of the assay-field data columns.
 }

 \usage{\method{getAssayFields}{BaseFileSpots}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:getColumns.BaseFileSpots]{*getColumns}()}
   For more information see \code{\link{BaseFileSpots}}.
 }
\keyword{internal}
\keyword{methods}
