%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getColumns.BaseFileSection}
\alias{getColumns.BaseFileSection}
\alias{BaseFileSection.getColumns}
\alias{getColumns.BaseFileSection}
\alias{getColumns,BaseFileSection-method}


 \title{Gets the column names of the data table}

 \description{
   Gets the column names of the data table, if it exists.
 }

 \usage{\method{getColumns}{BaseFileSection}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string \code{\link[base]{vector}}, if data table exists, otherwise
   \code{\link[base]{NULL}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
