%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSpots.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getColumns.BaseFileSpots}
\alias{getColumns.BaseFileSpots}
\alias{BaseFileSpots.getColumns}
\alias{getColumns.BaseFileSpots}
\alias{getColumns,BaseFileSpots-method}


 \title{Gets the names of the non-assay-field data columns}

 \description{
   Gets the names of the non-assay-field data columns.
 }

 \usage{\method{getColumns}{BaseFileSpots}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:getAssayFields.BaseFileSpots]{*getAssayFields}()}
   For more information see \code{\link{BaseFileSpots}}.
 }
\keyword{internal}
\keyword{methods}
