%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getCount.BaseFileAssays}
\alias{getCount.BaseFileAssays}
\alias{BaseFileAssays.getCount}
\alias{getCount.BaseFileAssays}
\alias{getCount,BaseFileAssays-method}


 \title{Gets the (expected) number of assays in the data table}

 \description{
   Gets the (expected) number of assays in the data table.
 }

 \usage{\method{getCount}{BaseFileAssays}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}, or \code{\link[base]{NULL}} if count is not specified.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:setCount.BaseFileAssays]{*setCount}()}
   \code{\link[aroma.Base:nbrOfAssays.BaseFileAssays]{*nbrOfAssays}()}
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
