%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getData.BaseFileSection}
\alias{getData.BaseFileSection}
\alias{BaseFileSection.getData}
\alias{getData.BaseFileSection}
\alias{getData,BaseFileSection-method}


 \title{Gets all or a subset of data of a BASE file section}

 \description{
   Gets all or a subset of data of a BASE file section.
 }

 \usage{\method{getData}{BaseFileSection}(this, fields=NULL, ...)}

 \arguments{
  \item{fields}{A \code{\link[base]{character}} \code{\link[base]{vector}} of field names to be retrieved.
     If \code{\link[base]{NULL}}, all fields are returned.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{data.frame}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
