%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getDataByIds.BaseFileAssays}
\alias{getDataByIds.BaseFileAssays}
\alias{BaseFileAssays.getDataByIds}
\alias{getDataByIds.BaseFileAssays}
\alias{getDataByIds,BaseFileAssays-method}


 \title{Gets a subset of data for one or all bioassays in the bioassay set}

 \description{
   Gets a subset of data for one or all bioassays in the bioassay set.
 }

 \usage{\method{getDataByIds}{BaseFileAssays}(this, ids, ...)}

 \arguments{
  \item{ids}{Ids of assays to be returned.}
  \item{...}{Arguments passed to
              \code{\link[getDataByKeys.BaseFileSection]{getDataByKeys()}}.}
 }

 \value{
   Returns a \code{\link[base]{data.frame}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
