%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSpots.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getDataFiles.BaseFileSpots}
\alias{getDataFiles.BaseFileSpots}
\alias{BaseFileSpots.getDataFiles}
\alias{getDataFiles.BaseFileSpots}
\alias{getDataFiles,BaseFileSpots-method}


 \title{Gets the filenames where data of an assay-field section is stored}

 \description{
   Gets the filenames where data of an assay-field section is stored.
 }

 \usage{\method{getDataFiles}{BaseFileSpots}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} \code{\link[base]{vector}}, or \code{\link[base]{NULL}}, if no data is stored on file.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:setDataFiles.BaseFileSpots]{*setDataFiles}()}
   \code{\link[aroma.Base:hasDataFiles.BaseFileSpots]{*hasDataFiles}()}
   For more information see \code{\link{BaseFileSpots}}.
 }
\keyword{internal}
\keyword{methods}
