%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getHeader.BaseFileSection}
\alias{getHeader.BaseFileSection}
\alias{BaseFileSection.getHeader}
\alias{getHeader.BaseFileSection}
\alias{getHeader,BaseFileSection-method}


 \title{Gets one header of a BASE section}

 \description{
   Gets one header of a BASE section.
 }

 \usage{\method{getHeader}{BaseFileSection}(this, name, ...)}

 \arguments{
  \item{name}{A \code{\link[base]{character}} string specifying the header to retrieve.}
  \item{...}{Arguments passed to \code{\link[aroma.Base:getHeaders.BaseFileSection]{*getHeaders}()}.}
 }

 \value{
   Returns the value of a header.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:getHeaders.BaseFileSection]{*getHeaders}()}.
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
