%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getId.BaseFileAssays}
\alias{getId.BaseFileAssays}
\alias{BaseFileAssays.getId}
\alias{getId.BaseFileAssays}
\alias{getId,BaseFileAssays-method}


 \title{Gets the id of all bioassays in the bioassay set}

 \description{
   Gets the id of all bioassays in the bioassay set.
 }

 \usage{\method{getId}{BaseFileAssays}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
