%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileBioassaysetinfo.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getName.BaseFileBioassaysetinfo}
\alias{getName.BaseFileBioassaysetinfo}
\alias{BaseFileBioassaysetinfo.getName}
\alias{getName.BaseFileBioassaysetinfo}
\alias{getName,BaseFileBioassaysetinfo-method}


 \title{Gets the name of the bioassay set}

 \description{
   Gets the name of the bioassay set.
 }

 \usage{\method{getName}{BaseFileBioassaysetinfo}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileBioassaysetinfo}}.
 }
\keyword{internal}
\keyword{methods}
