%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getParents.BaseFileAssays}
\alias{getParents.BaseFileAssays}
\alias{BaseFileAssays.getParents}
\alias{getParents.BaseFileAssays}
\alias{getParents,BaseFileAssays-method}


 \title{Gets the ids of the assay parents of all bioassays in the bioassay set}

 \description{
   Gets the ids of the assay parents of all bioassays in the bioassay set.
 }

 \usage{\method{getParents}{BaseFileAssays}(this, ids=NULL, ...)}

 \arguments{
  \item{ids}{Ids of assays to be returned. If \code{\link[base]{NULL}}, all are returned.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
