%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileParameters.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getPluginVersion.BaseFileParameters}
\alias{getPluginVersion.BaseFileParameters}
\alias{BaseFileParameters.getPluginVersion}
\alias{getPluginVersion.BaseFileParameters}
\alias{getPluginVersion,BaseFileParameters-method}


 \title{Gets the plugin version of a BASE parameters section}

 \description{
   Gets the plugin version of a BASE parameters section.
   This parameter is automatically passed by BASE for transformation
   plugins.
 }

 \usage{\method{getPluginVersion}{BaseFileParameters}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string, or \code{\link[base]{NULL}}, if missing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileParameters}}.
 }
\keyword{internal}
\keyword{methods}
