%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{hasColumns.BaseFileSection}
\alias{hasColumns.BaseFileSection}
\alias{BaseFileSection.hasColumns}
\alias{hasColumns.BaseFileSection}
\alias{hasColumns,BaseFileSection-method}


 \title{Checks if certain columns exists or not}

 \description{
   Checks if certain columns exists or not.
 }

 \usage{\method{hasColumns}{BaseFileSection}(this, names, ...)}

 \arguments{
  \item{names}{A \code{\link[base]{character}} \code{\link[base]{vector}} of length one or more.}
  \item{...}{Not used.}
 }

 \value{
   Returns \code{\link[base:logical]{TRUE}} for every column that exists, otherwise \code{\link[base:logical]{FALSE}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
