%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{hasData.BaseFileSection}
\alias{hasData.BaseFileSection}
\alias{BaseFileSection.hasData}
\alias{hasData.BaseFileSection}
\alias{hasData,BaseFileSection-method}


 \title{Checks if section has a data table}

 \description{
   Checks if section has a data table.
 }

 \usage{\method{hasData}{BaseFileSection}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns \code{\link[base:logical]{TRUE}} if a data table exists, otherwise \code{\link[base:logical]{FALSE}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
