%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSpots.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{hasDataFiles.BaseFileSpots}
\alias{hasDataFiles.BaseFileSpots}
\alias{BaseFileSpots.hasDataFiles}
\alias{hasDataFiles.BaseFileSpots}
\alias{hasDataFiles,BaseFileSpots-method}


 \title{Checks if data of an assay-field section is stored on file}

 \description{
   Checks if data of an assay-field section is stored on file.
 }

 \usage{\method{hasDataFiles}{BaseFileSpots}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{logical}} value.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:getDataFiles.BaseFileSpots]{*getDataFiles}()}
   \code{\link[aroma.Base:setDataFiles.BaseFileSpots]{*setDataFiles}()}
   For more information see \code{\link{BaseFileSpots}}.
 }
\keyword{internal}
\keyword{methods}
