%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{hasHeaders.BaseFileSection}
\alias{hasHeaders.BaseFileSection}
\alias{BaseFileSection.hasHeaders}
\alias{hasHeaders.BaseFileSection}
\alias{hasHeaders,BaseFileSection-method}


 \title{Checks if specified headers exists in a BASE section}

 \description{
   Checks if specified headers exists in a BASE section.
 }

 \usage{\method{hasHeaders}{BaseFileSection}(this, ...)}

 \arguments{
  \item{...}{Arguments passed to \code{\link[aroma.Base:getHeaders.BaseFileSection]{*getHeaders}()}.}
 }

 \value{
   Returns a \code{\link[base]{logical}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:getHeaders.BaseFileSection]{*getHeaders}()}.
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
