%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{hasSerialBioAssaySets.BaseFile}
\alias{hasSerialBioAssaySets.BaseFile}
\alias{BaseFile.hasSerialBioAssaySets}
\alias{hasSerialBioAssaySets.BaseFile}
\alias{hasSerialBioAssaySets,BaseFile-method}


 \title{Checks if the bioassay sets have data in serial format}

 \description{
   Checks if the bioassay sets have data in serial format.
 }

 \usage{\method{hasSerialBioAssaySets}{BaseFile}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{logical}} value.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFile}}.
 }
\keyword{internal}
\keyword{methods}
