%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSpots.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfAssays.BaseFileSpots}
\alias{nbrOfAssays.BaseFileSpots}
\alias{BaseFileSpots.nbrOfAssays}
\alias{nbrOfAssays.BaseFileSpots}
\alias{nbrOfAssays,BaseFileSpots-method}


 \title{Gets the number of assays in this section}

 \description{
   Gets the number of assays in this section.
 }

 \usage{\method{nbrOfAssays}{BaseFileSpots}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSpots}}.
 }
\keyword{internal}
\keyword{methods}
