%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfHeaders.BaseFileSection}
\alias{nbrOfHeaders.BaseFileSection}
\alias{BaseFileSection.nbrOfHeaders}
\alias{nbrOfHeaders.BaseFileSection}
\alias{nbrOfHeaders,BaseFileSection-method}


 \title{Gets the number of headers in a BASE section}

 \description{
   Gets the number of headers in a BASE section.
 }

 \usage{\method{nbrOfHeaders}{BaseFileSection}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
