%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileParameters.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfParameters.BaseFileParameters}
\alias{nbrOfParameters.BaseFileParameters}
\alias{BaseFileParameters.nbrOfParameters}
\alias{nbrOfParameters.BaseFileParameters}
\alias{nbrOfParameters,BaseFileParameters-method}


 \title{Gets the number of parameters in a BASE parameters section}

 \description{
   Gets the number of parameters in a BASE parameters section.
 }

 \usage{\method{nbrOfParameters}{BaseFileParameters}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileParameters}}.
 }
\keyword{internal}
\keyword{methods}
