%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfSections.BaseFile}
\alias{nbrOfSections.BaseFile}
\alias{BaseFile.nbrOfSections}
\alias{nbrOfSections.BaseFile}
\alias{nbrOfSections,BaseFile-method}

 \alias{BaseFile.length}
\alias{length.BaseFile}
\alias{length,BaseFile-method}


 \title{Gets the number of sections in this BASE file structure}

 \description{
   Gets the number of sections in this BASE file structure.
 }

 \usage{\method{nbrOfSections}{BaseFile}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFile}}.
 }
\keyword{internal}
\keyword{methods}
