%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% The BasePluginDispatcher class
\alias{main}
\alias{BasePlugin}

% The BaseFile class
\alias{appendSection}
\alias{getAllDataFiles}
\alias{getSection}
\alias{getSections}
\alias{isSerial}
\alias{nbrOfSections}
\alias{read}
\alias{removeAllDataFiles}
\alias{removeSection}
\alias{replaceSection}
\alias{write}
\alias{write.default}

% The BaseFileSection class
\alias{equals}
\alias{print}
\alias{as.list}
\alias{validate}
\alias{getType}
\alias{setType}
\alias{nbrOfHeaders}
\alias{setHeader}
\alias{getHeader}
\alias{getHeaders}
\alias{hasHeaders}
\alias{addColumn}
\alias{getColumns}
\alias{hasColumns}
\alias{setColumns}
\alias{getAssayFields}
\alias{hasData}
\alias{getData}
\alias{getDataByKeys}
\alias{setData}
\alias{setDataFieldByKeys}
\alias{sizeOfData}

% The BaseFileParameters class
\alias{attachParameters}
\alias{getParameter}
\alias{getParameters}
\alias{nbrOfParameters}
\alias{getPluginVersion}

% The BaseFileAssays class
\alias{getCount}
\alias{getAnnotationColumns}
\alias{getDataByIds}
\alias{getId}
\alias{getName}
\alias{getParents}
\alias{setAnnotationColumns}
\alias{setCount}
\alias{setDataFieldByIds}
\alias{setName}
\alias{setParents}

% The BaseFileSpots class
\alias{getAssays}
\alias{setAssays}
\alias{nbrOfAssays}
\alias{hasDataFiles}
\alias{getDataFiles}
\alias{setDataFiles}
\alias{hasSerialBioAssaySets}
\alias{setDataFields}

% The BaseFilePlugin class
\alias{addEnumerationParameter}
\alias{addEnumerationParameter.BaseFilePlugin}
\alias{addParameter}
\alias{addParameter.BaseFilePlugin}
\alias{addUsedColumns}
\alias{addUsedColumns.BaseFilePlugin}
\alias{addUsedFields}
\alias{addUsedFields.BaseFilePlugin}
\alias{buildReports}
\alias{getBaseFileSections}
\alias{getBaseFileSections.BaseFilePlugin}
\alias{nbrOfParameters.BaseFilePlugin}
\alias{setDescription}
\alias{setDescription.BaseFilePlugin}
\alias{setExecName}
\alias{setExecName.BaseFilePlugin}
\alias{setGeneAverages}
\alias{setGeneAverages.BaseFilePlugin}
\alias{setMinChannels}
\alias{setMinChannels.BaseFilePlugin}
\alias{setSerialFormat}
\alias{setSerialFormat.BaseFilePlugin}
\alias{setUrl}
\alias{setUrl.BaseFilePlugin}
\alias{setUsedColumns}
\alias{setUsedColumns.BaseFilePlugin}
\alias{setUsedFields}
\alias{setUsedFields.BaseFilePlugin}
\alias{setVersion}
\alias{setVersion.BaseFilePlugin}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
