%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{print.BaseFileSection}
\alias{print.BaseFileSection}
\alias{BaseFileSection.print}
\alias{print.BaseFileSection}
\alias{print,BaseFileSection-method}


 \title{Prints a BASE file section to stdout}

 \description{
   Prints a BASE file section to stdout.
 }

 \usage{\method{print}{BaseFileSection}(x, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
