%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{removeAllDataFiles.BaseFile}
\alias{removeAllDataFiles.BaseFile}
\alias{BaseFile.removeAllDataFiles}
\alias{removeAllDataFiles.BaseFile}
\alias{removeAllDataFiles,BaseFile-method}


 \title{Removes all external data files used by a BASE file structure}

 \description{
   Removes all external data files used by a BASE file structure.
 }

 \usage{\method{removeAllDataFiles}{BaseFile}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns (invisibly) \code{\link[base]{character}} \code{\link[base]{vector}} of the (successfully)
   removed files.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFile}}.
 }
\keyword{internal}
\keyword{methods}
