%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{removeSection.BaseFile}
\alias{removeSection.BaseFile}
\alias{BaseFile.removeSection}
\alias{removeSection.BaseFile}
\alias{removeSection,BaseFile-method}


 \title{Removes a section from a BASE file structure}

 \description{
   Removes a section from a BASE file structure.
 }

 \usage{\method{removeSection}{BaseFile}(this, section, ...)}

 \arguments{
  \item{section}{A \code{\link{BaseFileSection}} to be removed.}
  \item{...}{Not used.}
 }

 \value{
   Returns (invisibly) the number of sections removed.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFile}}.
 }
\keyword{internal}
\keyword{methods}
