%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{seq.BaseFile}
\alias{seq.BaseFile}
\alias{BaseFile.seq}
\alias{seq.BaseFile}
\alias{seq,BaseFile-method}


 \title{Gets an index sequence for the sections in this BASE file structure}

 \description{
   Gets an index sequence for the sections in this BASE file structure.
 }

 \usage{\method{seq}{BaseFile}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:nbrOfSections.BaseFile]{*nbrOfSections}()}.
   For more information see \code{\link{BaseFile}}.
 }
\keyword{internal}
\keyword{methods}
