%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSpots.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setAssays.BaseFileSpots}
\alias{setAssays.BaseFileSpots}
\alias{BaseFileSpots.setAssays}
\alias{setAssays.BaseFileSpots}
\alias{setAssays,BaseFileSpots-method}


 \title{Sets the ids of the assays in this section}

 \description{
   Sets the ids of the assays in this section.
 }

 \usage{\method{setAssays}{BaseFileSpots}(this, ids, ...)}

 \arguments{
  \item{ids}{The assay ids.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string \code{\link[base]{vector}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSpots}}.
 }
\keyword{internal}
\keyword{methods}
