%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setColumns.BaseFileSection}
\alias{setColumns.BaseFileSection}
\alias{BaseFileSection.setColumns}
\alias{setColumns.BaseFileSection}
\alias{setColumns,BaseFileSection-method}


 \title{Sets the column names of the data table}

 \description{
   Sets the column names of the data table.
 }

 \usage{\method{setColumns}{BaseFileSection}(this, names, ...)}

 \arguments{
  \item{names}{A \code{\link[base]{character}} \code{\link[base]{vector}}.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
