%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setData.BaseFileAssays}
\alias{setData.BaseFileAssays}
\alias{BaseFileAssays.setData}
\alias{setData.BaseFileAssays}
\alias{setData,BaseFileAssays-method}


 \title{Sets data of a BASE file section}

 \description{
   Sets data of a BASE file section. This calls the samemethod in the superclass, and updates
   the headers 'columns' and 'count' accordingly.
 }

 \usage{\method{setData}{BaseFileAssays}(this, ...)}

 \arguments{
  \item{...}{Passed to
            \code{\link[aroma.Base:setData.BaseFileSection]{setData()}}.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
