%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setData.BaseFileSection}
\alias{setData.BaseFileSection}
\alias{BaseFileSection.setData}
\alias{setData.BaseFileSection}
\alias{setData,BaseFileSection-method}


 \title{Sets data of a BASE file section}

 \description{
   Sets data of a BASE file section.
 }

 \usage{\method{setData}{BaseFileSection}(this, data, ...)}

 \arguments{
  \item{data}{A \code{\link[base]{list}} structure.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
