%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setDataFieldByIds.BaseFileAssays}
\alias{setDataFieldByIds.BaseFileAssays}
\alias{BaseFileAssays.setDataFieldByIds}
\alias{setDataFieldByIds.BaseFileAssays}
\alias{setDataFieldByIds,BaseFileAssays-method}


 \title{Sets a field of one or all bioassays in the bioassay set}

 \description{
   Sets a field of one or all bioassays in the bioassay set.
 }

 \usage{\method{setDataFieldByIds}{BaseFileAssays}(this, ids, field, values, ...)}

 \arguments{
  \item{...}{Arguments with names of assay ids and values to be assigned
     to the field.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
