%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSpots.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setDataFiles.BaseFileSpots}
\alias{setDataFiles.BaseFileSpots}
\alias{BaseFileSpots.setDataFiles}
\alias{setDataFiles.BaseFileSpots}
\alias{setDataFiles,BaseFileSpots-method}


 \title{Sets the filenames where data of an assay-field section is stored}

 \description{
   Sets the filenames where data of an assay-field section is stored.
 }

 \usage{\method{setDataFiles}{BaseFileSpots}(this, filenames, ...)}

 \arguments{
  \item{filenames}{A \code{\link[base]{character}} \code{\link[base]{vector}} of filenames.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[aroma.Base:getDataFiles.BaseFileSpots]{*getDataFiles}()}
   \code{\link[aroma.Base:hasDataFiles.BaseFileSpots]{*hasDataFiles}()}
   For more information see \code{\link{BaseFileSpots}}.
 }
\keyword{internal}
\keyword{methods}
