%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileAssays.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setName.BaseFileAssays}
\alias{setName.BaseFileAssays}
\alias{BaseFileAssays.setName}
\alias{setName.BaseFileAssays}
\alias{setName,BaseFileAssays-method}


 \title{Sets the names of one or all bioassays in the bioassay set}

 \description{
   Sets the names of one or all bioassays in the bioassay set.
 }

 \usage{\method{setName}{BaseFileAssays}(this, ...)}

 \arguments{
  \item{...}{Arguments with names of assay ids and values that are
     single \code{\link[base]{character}} strings of the corresponding assay names.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileAssays}}.
 }
\keyword{internal}
\keyword{methods}
