%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileSection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setType.BaseFileSection}
\alias{setType.BaseFileSection}
\alias{BaseFileSection.setType}
\alias{setType.BaseFileSection}
\alias{setType,BaseFileSection-method}


 \title{Sets the type of a BASE section}

 \description{
   Sets the type of a BASE section.
 }

 \usage{\method{setType}{BaseFileSection}(this, type, ...)}

 \arguments{
  \item{type}{A \code{\link[base]{character}} string specifying the new type.}
  \item{...}{Not used.}
 }

 \value{
   Returns (invisibly) the old type as a \code{\link[base]{character}} string.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileSection}}.
 }
\keyword{internal}
\keyword{methods}
