%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BaseFileBioassaysetinfo.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{validate.BaseFileBioassaysetinfo}
\alias{validate.BaseFileBioassaysetinfo}
\alias{BaseFileBioassaysetinfo.validate}
\alias{validate.BaseFileBioassaysetinfo}
\alias{validate,BaseFileBioassaysetinfo-method}


 \title{Validates bioassay-set information section}

 \description{
   Validates bioassay-set information section.
 }

 \usage{\method{validate}{BaseFileBioassaysetinfo}(this, action=c("error", "warning", "quiet"), ...)}

 \arguments{
  \item{action}{A \code{\link[base]{character}} string specifying action if a problem is
               detected. See \code{\link{validate.BaseFileSection}}() for details.}
  \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string explaining problems detected. If no problems
   where detected, \code{\link[base]{NULL}} is returned.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{BaseFileBioassaysetinfo}}.
 }
\keyword{internal}
\keyword{methods}
